# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Gestor de Pedidos Linamar API'
  description: 'Contrato HTTP del backend separado para la gestión de pedidos.'
  contact:
    email: soporte@linamar.local
  version: 0.1.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'Entorno local Docker'
paths:
...
  '/api/v1/campaigns/{campaign}/users':
    get:
      tags:
        - 'Usuarios de campaÃ±a'
      summary: 'Listar usuarios asignados a una campaÃ±a'
      operationId: listCampaignUsers
      responses:
        200:
          description: 'Usuarios obtenidos'
    post:
      tags:
        - 'Usuarios de campaÃ±a'
      summary: 'Asignar usuario a una campaÃ±a'
      operationId: assignCampaignUser
      responses:
        201:
          description: 'Usuario asignado'
        409:
          description: 'AsignaciÃ³n duplicada'
...
  /api/v1/orders:
    get:
      tags:
        - Pedidos
      summary: 'Listar pedidos'
      operationId: listOrders
      responses:
        200:
          description: 'Pedidos obtenidos'
    post:
      tags:
        - Pedidos
      summary: 'Crear pedido'
      operationId: createOrder
      responses:
        201:
          description: 'Pedido creado'
        200:
          description: 'Pedido ya existente por idempotencia'
...
components:
...
  securitySchemes:
    sanctum:
      type: http
      description: 'Token personal emitido por Laravel Sanctum'
      bearerFormat: token
      scheme: bearer
...
security:
  -
    sanctum: []